WIP: comparison tests not quite qworking yet
